# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: HTTP Conflict 409
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    SelfDescription:
      type: object
      properties:
        sdHash:
          type: string
        id: #!! Added id / credentialSubject
          type: string
        status:
          $ref: '#/components/schemas/SelfDescriptionStatus'
        issuer:
          type: string
        validatorDids:
          description: >-
            The credentialSubjects of the parties that validated (parts of) the
            Self-Description
          type: array
          items:
            type: string
        uploadDatetime:
          type: string
          format: date-time
          example: '2022-03-01T13:00:00Z'
          description: The time stamp (ISO8601) when the SD was uploaded.
        statusDatetime:
          type: string
          format: date-time
          example: '2022-05-11T15:30:00Z'
          description: >-
            The last time stamp (ISO8601) the status changed (for this
            Catalogue)
    SelfDescriptionStatus:
      type: string
      enum:
        - active
        - eol
        - deprecated
        - revoked
    SelfDescriptionResult:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/SelfDescription'
        content:
          type: string
...
    ResultPage:
      type: object
      properties:
        totalCount:
          type: integer
    SelfDescriptions:
      allOf:
        - $ref: '#/components/schemas/ResultPage'
        - type: object
          properties:
            items: 
              type: array
              items: 
                $ref: '#/components/schemas/SelfDescriptionResult'
...
  parameters:
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      required: false
      description: The number of items to skip before starting to collect the result set.
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      required: false
      description: The number of items to return.
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /self-descriptions:
    get:
      tags:
        - SelfDescriptions
      summary: Get the list of metadata of Self-Descriptions in the Catalogue
      operationId: readSelfDescriptions
      parameters: #!! Added filter parameters for all(?) metadata fields
        - name: uploadTimerange
          in: query
          description: >-
            Filter for the time range when the Self-Description was uploaded to
            the catalogue. The time range has to be specified as start time and
            end time as ISO8601 timestamp separated by a `/`.
          required: false
          example: 2022-03-01T13:00:00Z/2022-05-11T15:30:00Z
          schema:
            type: string
        - name: statusTimerange
          in: query
          description: >-
            Filter for the time range when the status of the Self-Description
            was last changed in the catalogue. The time range has to be
            specified as start time and end time as ISO8601 timestamp separated
            by a `/`.
          required: false
          example: 2022-03-01T13:00:00Z/2022-05-11T15:30:00Z
          schema:
            type: string
        - name: issuers
          in: query
          description: >-
            Filter for the issuer of the Self-Description. This is the unique ID
            of the Participant that has prepared the Self-Description.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false    
        - name: validators
          in: query
          description: >-
            Filter for a validator of the Self-Description. This is the unique
            ID of the Participant that validated (part of) the Self-Description.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false    
        - name: statuses
          in: query
          description: Filter for the status of the Self-Description.
          required: false
          schema:
            type: array
            items:
                $ref: '#/components/schemas/SelfDescriptionStatus'
          explode: false      
        - name: ids
          in: query
          description: Filter for a id/credentialSubject of the Self-Description.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false    
        - name: hashes
          in: query
          description: Filter for a hash of the Self-Description.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: withMeta
          in: query
          description: To add Self-Description Metadata to response or not.
          required: false
          schema:
            type: boolean
            default: true
        - name: withContent
          in: query
          description: To add Self-Description Content to response or not.
          required: false
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of meta data of available Self-Descriptions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfDescriptions'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - SelfDescriptions
      summary: Add a new SelfDescription to the catalogue
      operationId: addSelfDescription
      security:
        - jwt: []
      requestBody:
        description: The new SelfDescription
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        '201':
          description: Created
          headers:
            location:
              description: URL of the added Self-Description.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfDescription'
        '202':
          description: The request was accepted but the validation is not finished yet.
          headers:
            location:
              description: The URL of the status page.
              schema:
                type: string
            retry-after:
              description: >-
                The delay the client have to wait before requesting the status
                page.
              schema:
                type: integer
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /self-descriptions/{self_description_hash}:
    get:
      tags:
        - SelfDescriptions
      summary: >-
        Read a Self-Description by its hash. This returns the content of the
        self-description.
      description: Returns a single Self-Description
      operationId: readSelfDescriptionByHash
      parameters:
        - name: self_description_hash
          in: path
          description: Hash of the SelfDescription
          required: true
          schema:
            type: string
      security:
        - jwt: []
      responses:
        '200':
          description: The requested SelfDescription
          content:
            application/json:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - SelfDescriptions
      summary: Completely delete a self-description
      operationId: deleteSelfDescription
      parameters:
        - name: self_description_hash
          in: path
          description: Hash of the SelfDescription
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /self-descriptions/{self_description_hash}/revoke:
    post:
      tags:
        - SelfDescriptions
      summary: Change the lifecycle state of a SelfDescription to revoked.
      operationId: updateSelfDescription
      parameters:
        - name: self_description_hash
          in: path
          description: Hash of the SelfDescription
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Revoked
          headers:
            location:
              description: URL of the added SelfDescription
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfDescription'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...